+++ drift/lib/src/query_builder.dart
@@ -34,7 +34,9 @@ export 'query_builder/schema/view.dart';
 export 'query_builder/statements/delete.dart';
 export 'query_builder/statements/insert.dart';
 export 'query_builder/statements/select.dart';
+export 'query_builder/statements/statement.dart';
 export 'query_builder/statements/update.dart';
+export 'query_builder/statements/query.dart';
 
 export 'query_builder/results.dart';
 export 'query_builder/types.dart';